Refactor hybrid search into modular package
Extracted configuration, embedding, expansion, filtering, and Qdrant client logic from monolithic scripts into a new modular package under scripts/hybrid/. Updated corresponding legacy script files to use the new modules, improving maintainability and reusability.
