+++ scripts/hybrid/__init__.py
@@ -1,0 +1,15 @@
+"""
+Hybrid search package.
+
+Usage:
+    from scripts.hybrid import config, qdrant, embed, filters, expand, ranking
+    from scripts.hybrid.config import QDRANT_URL
+"""
+from scripts.hybrid import config
+from scripts.hybrid import qdrant
+from scripts.hybrid import embed
+from scripts.hybrid import filters
+from scripts.hybrid import expand
+from scripts.hybrid import ranking
+
+__all__ = ["config", "qdrant", "embed", "filters", "expand", "ranking"]
